pyproject.toml
[project]
name = "nonebot-plugin-sublike"
version = "0.1.0"
description = "插件描述"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "ByteColtX", email = "ByteColtX@gmail.com" }]
dependencies = [
  "nonebot-plugin-apscheduler>=0.5.0,<1.0.0", # 定时任务
  "nonebot-plugin-localstore>=0.7.4,<1.0.0",  # 存储文件
  "nonebot2>=2.4.0,<3.0.0",
  "nonebot-adapter-onebot>=2.4.0,<3.0.0",
]

[project.urls]
Homepage = "https://github.com/ByteColtX/nonebot-plugin-sublike"
Issues = "https://github.com/ByteColtX/nonebot-plugin-sublike/issues"
Repository = "https://github.com/ByteColtX/nonebot-plugin-sublike.git"

[dependency-groups]
dev = [
  "bump-my-version>=1.2.7",
  "ruff>=0.15.2,<1.0.0",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"


[tool.bumpversion]
current_version = "0.1.0"
commit = true
message = "release: bump vesion from {current_version} to {new_version}"
tag = true

[[tool.bumpversion.files]]
filename = "uv.lock"
search = "name = \"nonebot-plugin-sublike\"\nversion = \"{current_version}\""
replace = "name = \"nonebot-plugin-sublike\"\nversion = \"{new_version}\""


[tool.nonebot]
plugins = ["nonebot_plugin_sublike"]

[tool.ruff]
line-length = 88

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "I",     # isort
  "UP",    # pyupgrade
  "RUF",   # Ruff-specific rules
]
ignore = [
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
